update: unselect default si
