+++ tests/cypress/support/step_definitions/I select materials from dropdown.ts
@@ -1,6 +1,7 @@
-import { DataTable, When, Then } from "@badeball/cypress-cucumber-preprocessor";
-import MaterialDesignerPage from "../widgets/MaterialDesignerPage";
+import { DataTable, Then, When } from "@badeball/cypress-cucumber-preprocessor";
 import { parseTable } from "@mat3ra/tede/src/js/cypress/utils/table";
+
+import MaterialDesignerPage from "../widgets/MaterialDesignerPage";
 
 interface MaterialSelection {
     name: string;
@@ -10,19 +11,22 @@ interface MaterialSelection {
 When("I select materials from dropdown", (table: DataTable) => {
     const materials = parseTable<MaterialSelection>(table);
     const materialDesignerPage = new MaterialDesignerPage();
-    
     // Select each material from the dropdown
     materials.forEach(({ index }) => {
-        materialDesignerPage.designerWidget.jupyterLiteTransformationDialog.selectMaterialsIn(index);
+        materialDesignerPage.designerWidget.jupyterLiteTransformationDialog.selectMaterialsIn(
+            index,
+        );
     });
 });
 
 Then("I see materials selected", (table: DataTable) => {
     const materials = parseTable<MaterialSelection>(table);
     const materialDesignerPage = new MaterialDesignerPage();
-    
+
     // Verify each material is selected
     materials.forEach(({ index }) => {
-        materialDesignerPage.designerWidget.jupyterLiteTransformationDialog.verifyMaterialsOut(index);
+        materialDesignerPage.designerWidget.jupyterLiteTransformationDialog.verifyMaterialsOut(
+            index,
+        );
     });
-}); 
+});
